PrimitiveObsessionSniffer.getSmellNames fixed

+++ nosejob/src/main/java/com/codingrangers/nosejob/models/IAnalyser.java
@@ -7,7 +7,7 @@ public interface IAnalyser {
 
 	String getSmellName();
 
-	int getTotalSmellSeverity();
+	float getTotalSmellSeverity();
 
 	List<ISmell> getSmellInstances();
 }
